requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.6; python_version >= '3.9'
aiohttp==3.11.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
attrs==25.1.0; python_version >= '3.8'
beautifulsoup4==4.13.3; python_full_version >= '3.7.0'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
frozenlist==1.5.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
jinja2==3.1.5; python_version >= '3.7'
jiter==0.8.2; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
langchain==0.3.19; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.18; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.40; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.3.7; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.6; python_version >= '3.9' and python_version < '4.0'
langsmith==0.3.11; python_version >= '3.9' and python_version < '4.0'
loguru==0.7.3; python_version >= '3.5' and python_version < '4.0'
lxml==5.3.1; python_version >= '3.6'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
narwhals==1.28.0; python_version >= '3.8'
numpy==2.2.3; python_version >= '3.10'
openai==1.65.1; python_version >= '3.8'
orjson==3.10.15; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
playwright==1.50.0; python_version >= '3.9'
pluggy==1.5.0; python_version >= '3.8'
propcache==0.3.0; python_version >= '3.9'
protobuf==5.29.3; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.8.1; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pyee==12.1.1; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
pytest==8.3.4; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-decouple==3.8
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.9.4; python_full_version >= '3.8.0'
rpds-py==0.23.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlalchemy==2.0.38; python_version >= '3.7'
streamlit==1.42.2; python_version >= '3.9' and python_full_version != '3.9.7'
tenacity==9.0.0; python_version >= '3.8'
tiktoken==0.9.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.4.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
watchdog==6.0.0; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
